+++ common/src/main/java/com/human/common/property/HumanProperty.java
@@ -1,0 +1,8 @@
+package com.human.common.property;
+
+import com.blib.api.common.property.v1.BLibPropertyKey;
+
+public record HumanProperty<T>(
+    BLibPropertyKey.Leaf<T> key,
+    T defaultValue
+) {}